Move config classes to config folder in infrastructure module
