+++ src/main/kotlin/no/nb/mlt/wls/infrastructure/config/ProxyConfig.kt
@@ -1,4 +1,4 @@
-package no.nb.mlt.wls.infrastructure
+package no.nb.mlt.wls.infrastructure.config
 
 import org.springframework.boot.context.properties.ConfigurationProperties
