+++ src/test/kotlin/no/nb/mlt/wls/ArchitectureTest.kt
@@ -1,6 +1,8 @@
 package no.nb.mlt.wls
 
+import com.tngtech.archunit.base.DescribedPredicate
 import com.tngtech.archunit.core.domain.JavaClasses
+import com.tngtech.archunit.core.domain.properties.HasName.Predicates.nameMatching
 import com.tngtech.archunit.junit.AnalyzeClasses
 import com.tngtech.archunit.junit.ArchTest
 import com.tngtech.archunit.lang.syntax.ArchRuleDefinition.noClasses
@@ -49,6 +51,7 @@ internal class ArchitectureTest {
             .matching("no.nb.mlt.wls.infrastructure..(*)")
             .should()
             .notDependOnEachOther()
+            .ignoreDependency(nameMatching(".*Config"), DescribedPredicate.alwaysTrue())
             .check(javaClasses)
     }
 }
